Change UsageFormatter into an interface
